feat(liquidity): Upgrate tabbar to v2
