+++ src/app.config.ts
@@ -1,6 +1,6 @@
 export default {
-  entryPagePath: 'views/BmpHome/index',
-  pages: ['views/BmpHome/index', 'views/webview'],
+  entryPagePath: 'views/Swap/bmp/index',
+  pages: ['views/Swap/bmp/index', 'views/bmp/liquidity/index', 'views/webview'],
   darkmode: true,
   window: {
     backgroundTextStyle: 'light',
@@ -9,4 +9,24 @@ export default {
     navigationBarTitleText: 'PancakeSwap',
   },
   themeLocation: './theme.json',
+  tabBar: {
+    color: '#7A6EAA',
+    selectedColor: '#7645D9',
+    backgroundColor: '#fff',
+    borderStyle: 'white',
+    list: [
+      {
+        iconPath: 'images/trade.png',
+        selectedIconPath: 'images/trade-select.png',
+        pagePath: 'views/Swap/bmp/index',
+        text: 'Exchange',
+      },
+      {
+        iconPath: 'images/earn.png',
+        selectedIconPath: 'images/earn-select.png',
+        pagePath: 'views/bmp/liquidity/index',
+        text: 'Liquidity',
+      },
+    ],
+  },
 }
