+++ src/views/RemoveLiquidity/index.bmp.tsx
@@ -9,7 +9,7 @@ import { Slider } from '@binance/mp-components'
 import { Button, Text, AddIcon, ArrowDownIcon, CardBody, Box, Flex, useModal } from '@pancakeswap/uikit'
 import { BigNumber } from '@ethersproject/bignumber'
 import { useTranslation } from 'contexts/Localization'
-import { useLiquidity, LiquidityPage } from 'views/BmpHome/context/swapContext.bmp'
+import { useLiquidity, LiquidityPage } from 'views/bmp/liquidity/liquidityContext'
 import { AutoColumn, ColumnCenter } from '../../components/Layout/Column'
 import TransactionConfirmationModal, { ConfirmationModalContent } from '../../components/TransactionConfirmationModal'
 import CurrencyInputPanel from '../../components/CurrencyInputPanel'
