+++ src/views/bmp/BmpPage/components/farmsWrapper.bmp.tsx
