+++ src/views/bmp/BmpPage/constants.ts
