+++ src/views/bmp/liquidity/index.config.ts
